Move FrontendInputs into separate .h and .cpp files.
